requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
authentication==1.1.0
backports.zoneinfo==0.2.1
blinker==1.6.2
click==8.1.6
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Flask==2.3.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
pycryptodome==3.18.0
PyJWT==2.8.0
pytz==2023.3
redis==4.6.0
sqlparse==0.4.4
typing_extensions==4.7.1
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.16.2
